core: new diskio library
Beginnings of a diskio library capable of reading and writing, and
with proper handling of bounce buffers and segment boundaries.

+++ core/include/disk.h
@@ -1,11 +1,13 @@
 #ifndef DISK_H
 #define DISK_H
 
+#include <stdint.h>
+
 #define SECTOR_SHIFT     9
 #define SECTOR_SIZE      (1 << SECTOR_SHIFT)
 
-typedef unsigned int sector_t;
-typedef unsigned int block_t;
+typedef uint64_t sector_t;
+typedef uint32_t block_t;
 
 extern void read_sectors(char *, sector_t, int);
 extern void getoneblk(char *, block_t, int);
